+++ src/jsv_type_date.erl
@@ -17,7 +17,7 @@
 -behaviour(jsv_type).
 
 -export([verify_constraint/2, format_constraint_violation/2,
-         validate_type/1, validate_constraint/4, canonicalize/3, generate/2,
+         validate_type/2, validate_constraint/4, canonicalize/3, generate/2,
          format_date/1, is_valid_date/1]).
 
 -export_type([constraint/0]).
@@ -52,9 +52,14 @@ format_constraint_violation({min, Min}, _) ->
 format_constraint_violation({max, Max}, _) ->
   {"value must be lower or equal to ~s", [format_date(Max)]}.
 
-validate_type(Value) when is_binary(Value) ->
-  parse_date(Value);
-validate_type(_) ->
+validate_type(Value, _) when is_binary(Value) ->
+  case parse_date(Value) of
+    {ok, Date} ->
+      {ok, Value, Date};
+    error ->
+      error
+  end;
+validate_type(_, _) ->
   error.
 
 validate_constraint(_, {min, Min}, Date, _) ->