+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 - Add experimental support for an extra set of data in type definitions. Use
   it for generation and validation functions.
 - Add a `unique_elements` constraint for arrays.
-- Add an `invalid_member_handling` option to treat object with a `members`
-  constraint as records and define how to handle members which are not part of
-  it.
+- Add an `invalid_member_handling` option to control how to handle object
+  members which are not part of the `members` constraint.
 - Check object member validity during generation
+- Add a `null_member_handling` option to control whether to keep or remove
+  null object members, which is useful for systems which consider a null value
+  equivalent to the absence of a member.
 - Add `jsv:verify_definition/1`.
 - Introduce the `{one_of, Definitions}` definition combinator.
 ## Bugs